+++ examples/data-sources/keel_permit/main.tf
@@ -1,4 +1,4 @@
 data "keel_permit" "last_denied" {
-  decision   = "deny"
-  limit      = 1
+  decision = "deny"
+  limit    = 1
 }